app-config-web/pyproject.toml
[project]
name = "azure-app-config-reader"
version = "0.1.0"
description = "Azure App Configuration reader using Azure Identity"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.15.0",
    "azure-appconfiguration>=1.5.0",
    "flask>=3.0.0",
]

[project.scripts]
app-config-reader = "azure_app_config:main"
